NewTeleOp_22.06.2018
I added a new driving system that is similar to this year's mecha falcons robot system. The horizontal motor is called strafeDrive. I also added a new method called servoContinousMotion. This method allows for us to control the main servo using the joysticks. I also made the default speed 40% but I added a turbo option. The turbo option is accessed by clicking and holding the left joystick and runs at 90% speed. I also condensed the telemetry data code in the loop method to give it a clearer overview. I am also starting to work on two different methods called fipAndDrop and automatedServoPlacement. The flipAndDrop method will with the press of a button place flip the cubes and place them on the robot. The automatedServoPlacement method is explained in the comments in the code.